Repair differences
