+++ README.md
@@ -115,7 +115,19 @@ npm install
 
 3. Paste the content you copied from `.env.template` in `.env`.
 
-4. Fill the keys with the secret values.
+4. Fill the keys with the secret values. For example:
+
+```
+DB=
+DBUSER=
+DBPASSWORD=
+```
+
+To access variables use `process.env.DB_HOST` for example.
+
+This is a shell file so there can't be spaces around `=`
+
+For more information refer to: https://www.npmjs.com/package/dotenv
 
 # Running the Project
 
@@ -152,17 +164,3 @@ List of commands:
 - `dev`: Start the application with nodemon
 - `lint`: Run eslint
 - `lint-fix`: Run eslint and let it fix the errors
-
-# Filling out .env file
-
-Add environment-specific variables on new lines in the form of NAME=VALUE. For example:
-```
-DB_HOST=localhost
-DB_USER=root
-DB_PASS=s1mpl3
-```
-To access variables use ```process.env.DB_HOST``` for example. 
-
-This is a shell file so there can't be spaces around = 
-
-For more information refer to: https://www.npmjs.com/package/dotenv
