--- service/.env.example
@@ -1,2 +1,0 @@
-MAVENLINK_TASK_URL="https://api.mavenlink.com/api/v1/stories"
-MAVENLINK_TOKEN="{Add your token}"
