+++ service/.env.template
@@ -1,10 +1,14 @@
 # App port
-port=
+PORT=
 
 # Database
-db=
-dbUser=
-dbPassword=
+DB=
+DBUSER=
+DBPASSWORD=
 
 # Used for encoding and decoding JSON Web Tokens
-jwtSecret=
+JWTSECRET=
+
+# Mavenlink
+MAVENLINK_TASK_URL="https://api.mavenlink.com/api/v1/stories"
+MAVENLINK_TOKEN=
