+++ service/src/routes/index.ts
@@ -1,7 +1,11 @@
+import express from 'express'
+
 import task from './task'
 import user from './user';
 
-export default {
-  task,
-  user
-};
+const routes = (app: express.Application) => {
+  app.use('/', user)
+  app.use('/tasks', task)
+}
+
+export default routes;
